Added One Time Sounds Prefab
Created a prefab to play sounds only one time through a simple script. It is used when we want to play a sound when a certain event happens or the user enters or leaves a specific area ..etc. However, there's a SerializeField that could disable the one time playback effect and make the sound play every time the player collides with the sound source (There's a pause time between every playback that can also be set by preference from the inspector).