Applied Pause Animation to the Enemy
Used the pause animation (Looking Around) and scripted it to play at specific time in the patrol route I created for the enemy.
For example, the enemy is patrolling the church, he is going from point A to point B. Between point A and point B the animation "Walking with Axe" will play, but, upon reaching point A OR point B the animation "Looking Around" will play and the cycle repeats itself.