Completed the locked door system
Now, some doors will be locked and will require a specific key to unlock. There are three keys in the game for locked doors, house, room, and cabin keys. If the player doesn't has said key, the game will inform him that he needs that specific key. If he does have the key then the door can be opened and closed.