Created Player Stamina System
Now the player can not run infinitely, the player has stamina that is decreased over time while running, if the stamina is below a certain amount, a breathing sound plays and loops. When the stamina is depleted, the player is forced to walk even if he/she is pressing down the Left Shift key. If the player stopped holding the Left Shift key and walked through W key then the stamina will start refilling again at a constant rate and the breathing sound will stop after a while.